# perf: improve performance of Link and routes