… packages/react-router/src/route.ts
… packages/react-router/tests/link.test-d.tsx
… packages/react-router/tests/route.test-d.tsx