### TS Performance Improvments

## Purpose

Type checking large code bases with many `Route`'s and `Link`'s can get quite sluggish. This PR introduces a number of performance optimizations I've discovered while profiling on the large file based example. The purpose of this PR is to improve suggestion times for such code bases and build times

## Extended Diagnostics

On main the following diagnostics are present for the example large file based

```
> tsc --extendedDiagnostics

Files:                         589
Lines of Library:            38118
Lines of Definitions:        77132
Lines of TypeScript:         12777
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                113478
Symbols:                    351283
Types:                       62925
Instantiations:            1541766
Memory used:               311542K
Assignability cache size:   113357
Identity cache size:          3707
Subtype cache size:              0
Strict subtype cache size:  162006
I/O Read time:               0.08s
Parse time:                  0.55s
ResolveModule time:          0.11s
ResolveTypeReference time:   0.01s
ResolveLibrary time:         0.02s
Program time:                0.84s
Bind time:                   0.24s
Check time:                  5.64s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  6.72s
```

After the changes introduced in this PR the instantiations and check time are reduced

```
> tsc --extendedDiagnostics

Files:                         589
Lines of Library:            38118
Lines of Definitions:        77133
Lines of TypeScript:         12777
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                113461
Symbols:                    189848
Types:                       50776
Instantiations:             592405
Memory used:               222842K
Assignability cache size:    34415
Identity cache size:          3718
Subtype cache size:              0
Strict subtype cache size:       0
I/O Read time:               0.07s
Parse time:                  0.54s
ResolveModule time:          0.11s
ResolveTypeReference time:   0.01s
ResolveLibrary time:         0.02s
Program time:                0.83s
Bind time:                   0.23s
Check time:                  3.82s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  4.88s
```

## Profiling

This is the profile for the large file route on main.

![image](https://github.com/TanStack/router/assets/7883601/542a6399-1b33-41ab-952e-a8b35ba67ead)

Going deep into each wall of the profile. The biggest wall is the route tree definition. There is a number of things to improve here.

- Performance of each individual file route types
- Variance checks on each Route
- Type parameter constraints on each Route
- Improving performance of merging properties from each parent route
- Only `Expand` when necessary and not eagerly

![image](https://github.com/TanStack/router/assets/7883601/3c73a3c3-1f2e-416c-a85c-f4aace9fd88a)

And the next significant thing is each `Link`. Each `Link` can be improved by the following:

- Separating `LinkComponentProps` to cache props for a component passed to `createLink`
- Use the resolved route for relative paths to narrow down the union of paths relevant for autocomplete
- Use `CheckPath` directly on the `to` prop to avoid an intersection
- Evaluate less of `params` and `search` until the user use these props

![image](https://github.com/TanStack/router/assets/7883601/a7b84d57-e0c6-437c-b0a2-cced0c9ddc1a)

After the performance improvements the route tree looks more like this

![image](https://github.com/TanStack/router/assets/7883601/86a2626b-1776-4eff-8cf8-eddde4cffc1e)

and each `Link` looks like this

![image](https://github.com/TanStack/router/assets/7883601/6d3a9c3c-d9b3-408f-9783-aaca02f7f583)

`Link` may seem like a small improvement for one `Link` but when you have many in large code bases this quickly accumulates